prefect_flows/requirements.txt
prefect
supabase
python-dotenv
pydantic
elasticsearch>=8.0.0
dateparser
